+++ examples/client-ruby.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+set -o nounset
+
+cd ../
+make build-ruby
+cd examples/
+
+ruby usageruby.rb